% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.R
\name{glance.rdrobust}
\alias{glance.rdrobust}
\title{RDD extensions for glance}
\usage{
\method{glance}{rdrobust}(x, ...)
}
\arguments{
\item{x}{An \code{rdd} object created by the \code{rdrobust} package.}

\item{...}{Additional arguments passed to \code{modelsummary()}.}
}
\description{
Glance methods for \code{rdd} objects created by the \code{rdrobust} package.
}
