% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.R
\name{tidy.rdrobust}
\alias{tidy.rdrobust}
\title{RDD extensions for tidy}
\usage{
\method{tidy}{rdrobust}(x, ...)
}
\arguments{
\item{x}{An \code{rdd} object created by the \code{rdrobust} package.}

\item{...}{Additional arguments passed to \code{modelsummary()}.}
}
\description{
Tidy methods for \code{rdd} objects created by the \code{rdrobust} package.
}
\examples{
\dontrun{
x <- runif(100, -10, 10)
y <- 10 + 2*x + 2*(x>=0) + rnorm(100)
mod <- rdrobust(y,x)

tidy(mod)
glance(mod)
}

}
