% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.R
\name{tidy.rdrobust}
\alias{tidy.rdrobust}
\title{RDD extensions for model summary}
\usage{
tidy.rdrobust(model, ...)
}
\arguments{
\item{model}{An \code{rdd} object created by the \code{rdrobust} package.}

\item{...}{Additional arguments passed to \code{modelsummary()}.}
}
\description{
These functions provide extensions to the \code{modelsummary()} function for
\code{rdd} objects created by the \code{rdrobust} package.
}
